build(web-app): automate Firebase service worker generation using environment variables and update deployment workflows
