+++ .env.example
@@ -16,14 +16,14 @@ NEXT_PUBLIC_FIREBASE_VAPID_KEY=
 API_URL=
 APP_PORT=
 
-# Backend
+# Database
 DB_USER=
 DB_PASSWORD=
 DB_NAME=
 DB_HOST=
 DB_PORT=
-DATABASE_URL=
 
+#Backend
 DATABASE_URL=
 SERVER_PORT=
 NODE_ENV=
