+++ web-app/package.json
@@ -2,7 +2,9 @@
   "name": "@vigilart/web-app",
   "version": "0.1.0",
   "scripts": {
+    "predev": "dotenv -e ../.env -- node scripts/generate-sw.js",
     "dev": "dotenv -e ../.env -- next dev --turbopack",
+    "prebuild": "node scripts/generate-sw.js",
     "build": "next build --turbopack",
     "start": "next start"
   },
